requirements.txt
Flask>=2.3
PyMuPDF>=1.26
PyPDF2>=3.0
fonttools>=4.38
pytesseract>=0.3.10
Pillow>=9.5
